requirements.txt
Flask
flask-cors
gunicorn
psycopg2-binary
python-dotenv
requests
bcrypt
PyJWT